Refactor data processing scripts and update schema
- Changed TIPO from 'openai' to 'huggingface' in crea_file_json_huggingface.py.
- Reduced VALIDATION_SIZE from 0.05 to 0.006 in pulisci_splitta_dati.py.
- Updated STRATIFY_COLUMNS to include additional columns in pulisci_splitta_dati.py.
- Corrected column names for target_columns in pulisci_splitta_dati.py.
- Changed data type of distanza_oai from int to float in schema_json.py for both ReportData and ReportDataOneLevel classes.

+++ utils/schema_json.py
@@ -113,7 +113,7 @@ class ReportData(BaseModel):
     posizione: Posizione | None
     spessore_parietale: int | None
     estensione_cranio_caudale: int | None
-    distanza_oai: int | None
+    distanza_oai: float | None
     riflessione_peritoneale_anteriore: RiflessionePeritonealeAnteriore | None
     infiltrazione_tessuto_adiposo: InfiltrazioneTessutoAdiposo | None
     infiltrazione_sfinteri: InfiltrazioneSfinteri | None
@@ -144,7 +144,7 @@ class ReportDataOneLevel(BaseModel):
     posizione: Posizione | None
     spessore_parietale: int | None
     estensione_cranio_caudale: int | None
-    distanza_oai: int | None
+    distanza_oai: float | None
     riflessione_peritoneale_anteriore: RiflessionePeritonealeAnteriore | None
     infiltrazione_tessuto_adiposo: InfiltrazioneTessutoAdiposo | None
     infiltrazione_sfinteri: InfiltrazioneSfinteri | None
